requirements.txt
streamlit>=1.30.0
SpeechRecognition>=3.10.0
openai-whisper>=20231117
pydub>=0.25.1
imageio-ffmpeg>=0.4.8
torch>=2.0.0
python-docx>=1.0.1
pathlib>=1.0.1
typing>=3.7.4.3